requirements.txt
Django==2.2
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.2.1
mysqlclient==1.4.6
pytz==2019.3
sqlparse==0.3.1
